회원 인가 기능 구현

+++ src/main/java/com/evoting/controller/dto/AgendaDto.java
@@ -14,7 +14,7 @@ public class AgendaDto {
     private String contents;
     private AgendaStatus agendaStatus;
     private AgendaType agendaType;
-    private Long maxCount;
+    private Integer maxCount;
 
     public AgendaDto(Agenda agenda) {
         agendaId = agenda.getId();
